from home

+++ aws/resource_aws_subnet_test.go
@@ -18,25 +18,26 @@ func init() {
 	resource.AddTestSweepers("aws_subnet", &resource.Sweeper{
 		Name: "aws_subnet",
 		F:    testSweepSubnets,
-		// When implemented, these should be moved to aws_network_interface
-		// and aws_network_interface set as dependency here.
 		Dependencies: []string{
 			"aws_autoscaling_group",
 			"aws_batch_compute_environment",
 			"aws_beanstalk_environment",
 			"aws_db_instance",
 			"aws_directory_service_directory",
+			"aws_ec2_transit_gateway_vpc_attachment",
 			"aws_eks_cluster",
 			"aws_elasticache_cluster",
 			"aws_elasticache_replication_group",
 			"aws_elasticsearch_domain",
 			"aws_elb",
-			"aws_instance",
+			"aws_emr_cluster",
 			"aws_lambda_function",
 			"aws_lb",
 			"aws_mq_broker",
+			"aws_network_interface",
 			"aws_redshift_cluster",
 			"aws_spot_fleet_request",
+			"aws_vpc_endpoint",
 		},
 	})
 }
@@ -53,6 +54,7 @@ func testSweepSubnets(region string) error {
 			{
 				Name: aws.String("tag-value"),
 				Values: []*string{
+					aws.String("terraform-testacc-*"),
 					aws.String("tf-acc-*"),
 				},
 			},
@@ -73,6 +75,14 @@ func testSweepSubnets(region string) error {
 	}
 
 	for _, subnet := range resp.Subnets {
+		if subnet == nil {
+			continue
+		}
+
+		if aws.BoolValue(subnet.DefaultForAz) {
+			continue
+		}
+
 		// delete the subnet
 		_, err := conn.DeleteSubnet(&ec2.DeleteSubnetInput{
 			SubnetId: subnet.SubnetId,
@@ -112,12 +122,12 @@ func TestAccAWSSubnet_basic(t *testing.T) {
 	var v ec2.Subnet
 
 	testCheck := func(*terraform.State) error {
-		if *v.CidrBlock != "10.1.1.0/24" {
-			return fmt.Errorf("bad cidr: %s", *v.CidrBlock)
+		if aws.StringValue(v.CidrBlock) != "10.1.1.0/24" {
+			return fmt.Errorf("bad cidr: %s", aws.StringValue(v.CidrBlock))
 		}
 
-		if *v.MapPublicIpOnLaunch != true {
-			return fmt.Errorf("bad MapPublicIpOnLaunch: %t", *v.MapPublicIpOnLaunch)
+		if !aws.BoolValue(v.MapPublicIpOnLaunch) {
+			return fmt.Errorf("bad MapPublicIpOnLaunch: %t", aws.BoolValue(v.MapPublicIpOnLaunch))
 		}
 
 		return nil
@@ -142,6 +152,11 @@ func TestAccAWSSubnet_basic(t *testing.T) {
 						"aws_subnet.foo",
 						"arn",
 						regexp.MustCompile(`^arn:[^:]+:ec2:[^:]+:\d{12}:subnet/subnet-.+`)),
+					testAccCheckResourceAttrAccountID("aws_subnet.foo", "owner_id"),
+					resource.TestCheckResourceAttrSet(
+						"aws_subnet.foo", "availability_zone"),
+					resource.TestCheckResourceAttrSet(
+						"aws_subnet.foo", "availability_zone_id"),
 				),
 			},
 		},
@@ -213,14 +228,38 @@ func TestAccAWSSubnet_enableIpv6(t *testing.T) {
 	})
 }
 
+func TestAccAWSSubnet_availabilityZoneId(t *testing.T) {
+	var v ec2.Subnet
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:      func() { testAccPreCheck(t) },
+		IDRefreshName: "aws_subnet.foo",
+		Providers:     testAccProviders,
+		CheckDestroy:  testAccCheckSubnetDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccSubnetConfigAvailabilityZoneId,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckSubnetExists(
+						"aws_subnet.foo", &v),
+					resource.TestCheckResourceAttrSet(
+						"aws_subnet.foo", "availability_zone"),
+					resource.TestCheckResourceAttr(
+						"aws_subnet.foo", "availability_zone_id", "usw2-az3"),
+				),
+			},
+		},
+	})
+}
+
 func testAccCheckAwsSubnetIpv6BeforeUpdate(t *testing.T, subnet *ec2.Subnet) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		if subnet.Ipv6CidrBlockAssociationSet == nil {
 			return fmt.Errorf("Expected IPV6 CIDR Block Association")
 		}
 
-		if *subnet.AssignIpv6AddressOnCreation != true {
-			return fmt.Errorf("bad AssignIpv6AddressOnCreation: %t", *subnet.AssignIpv6AddressOnCreation)
+		if !aws.BoolValue(subnet.AssignIpv6AddressOnCreation) {
+			return fmt.Errorf("bad AssignIpv6AddressOnCreation: %t", aws.BoolValue(subnet.AssignIpv6AddressOnCreation))
 		}
 
 		return nil
@@ -229,8 +268,8 @@ func testAccCheckAwsSubnetIpv6BeforeUpdate(t *testing.T, subnet *ec2.Subnet) res
 
 func testAccCheckAwsSubnetIpv6AfterUpdate(t *testing.T, subnet *ec2.Subnet) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		if *subnet.AssignIpv6AddressOnCreation != false {
-			return fmt.Errorf("bad AssignIpv6AddressOnCreation: %t", *subnet.AssignIpv6AddressOnCreation)
+		if aws.BoolValue(subnet.AssignIpv6AddressOnCreation) {
+			return fmt.Errorf("bad AssignIpv6AddressOnCreation: %t", aws.BoolValue(subnet.AssignIpv6AddressOnCreation))
 		}
 
 		return nil
@@ -311,7 +350,7 @@ func testAccCheckSubnetExists(n string, v *ec2.Subnet) resource.TestCheckFunc {
 const testAccSubnetConfig = `
 resource "aws_vpc" "foo" {
 	cidr_block = "10.1.0.0/16"
-	tags {
+	tags = {
 		Name = "terraform-testacc-subnet"
 	}
 }
@@ -320,7 +359,7 @@ resource "aws_subnet" "foo" {
 	cidr_block = "10.1.1.0/24"
 	vpc_id = "${aws_vpc.foo.id}"
 	map_public_ip_on_launch = true
-	tags {
+	tags = {
 		Name = "tf-acc-subnet"
 	}
 }
@@ -330,7 +369,7 @@ const testAccSubnetConfigPreIpv6 = `
 resource "aws_vpc" "foo" {
 	cidr_block = "10.10.0.0/16"
 	assign_generated_ipv6_cidr_block = true
-	tags {
+	tags = {
 		Name = "terraform-testacc-subnet-ipv6"
 	}
 }
@@ -339,7 +378,7 @@ resource "aws_subnet" "foo" {
 	cidr_block = "10.10.1.0/24"
 	vpc_id = "${aws_vpc.foo.id}"
 	map_public_ip_on_launch = true
-	tags {
+	tags = {
 		Name = "tf-acc-subnet-ipv6"
 	}
 }
@@ -349,7 +388,7 @@ const testAccSubnetConfigIpv6 = `
 resource "aws_vpc" "foo" {
 	cidr_block = "10.10.0.0/16"
 	assign_generated_ipv6_cidr_block = true
-	tags {
+	tags = {
 		Name = "terraform-testacc-subnet-ipv6"
 	}
 }
@@ -360,7 +399,7 @@ resource "aws_subnet" "foo" {
 	ipv6_cidr_block = "${cidrsubnet(aws_vpc.foo.ipv6_cidr_block, 8, 1)}"
 	map_public_ip_on_launch = true
 	assign_ipv6_address_on_creation = true
-	tags {
+	tags = {
 		Name = "tf-acc-subnet-ipv6"
 	}
 }
@@ -370,7 +409,7 @@ const testAccSubnetConfigIpv6UpdateAssignIpv6OnCreation = `
 resource "aws_vpc" "foo" {
 	cidr_block = "10.10.0.0/16"
 	assign_generated_ipv6_cidr_block = true
-	tags {
+	tags = {
 		Name = "terraform-testacc-subnet-assign-ipv6-on-creation"
 	}
 }
@@ -381,7 +420,7 @@ resource "aws_subnet" "foo" {
 	ipv6_cidr_block = "${cidrsubnet(aws_vpc.foo.ipv6_cidr_block, 8, 1)}"
 	map_public_ip_on_launch = true
 	assign_ipv6_address_on_creation = false
-	tags {
+	tags = {
 		Name = "tf-acc-subnet-assign-ipv6-on-creation"
 	}
 }
@@ -391,7 +430,7 @@ const testAccSubnetConfigIpv6UpdateIpv6Cidr = `
 resource "aws_vpc" "foo" {
 	cidr_block = "10.10.0.0/16"
 	assign_generated_ipv6_cidr_block = true
-	tags {
+	tags = {
 		Name = "terraform-testacc-subnet-ipv6-update-cidr"
 	}
 }
@@ -402,8 +441,30 @@ resource "aws_subnet" "foo" {
 	ipv6_cidr_block = "${cidrsubnet(aws_vpc.foo.ipv6_cidr_block, 8, 3)}"
 	map_public_ip_on_launch = true
 	assign_ipv6_address_on_creation = false
-	tags {
+	tags = {
 		Name = "tf-acc-subnet-ipv6-update-cidr"
 	}
+}
+`
+
+const testAccSubnetConfigAvailabilityZoneId = `
+provider "aws" {
+  region = "us-west-2"
+}
+
+resource "aws_vpc" "foo" {
+  cidr_block = "10.1.0.0/16"
+  tags = {
+    Name = "terraform-testacc-subnet"
+  }
+}
+
+resource "aws_subnet" "foo" {
+  cidr_block = "10.1.1.0/24"
+  vpc_id = "${aws_vpc.foo.id}"
+  availability_zone_id = "usw2-az3"
+  tags = {
+    Name = "tf-acc-subnet"
+  }
 }
 `
